Add code to integrate data from count matricies
Get data through inital integration to further assess data quality.
NOTE: may need to go back and re-run due to the inability to
confidently identify which gene symbols coorespond to the MT
features that are to be used for intial QC filtering.

Will likely need to add a new formal to load10x() to allow for
the User to specify a list of MT features rather than use the
grep method. Actually, might be able to use regex syntax to
provide a list of features without modifying the function.
(i.e. "feat1|feat2")

JOBID: 5254755