Add DE and DA analysis to cleaned T cell subset
This is an over commit. Added:
 - code to clean and reintegrate the data to filter out  low quality cells
 - new cell type annotations
 - permutation testing for DA (both at sample and condition levels)
 - pseudobulk genes that define "T_stressed" cluster - poss a low Q cluster
 - DE analysis and plotting for all T cells and within each subtype
 - GSEA using the DE results

Current intpretation:
NK and CD8 T cells appear to be have th most profound responses, while the
cluster of "T_stressed" cells is overreprested in the diseased samples. Unsure
if the "T_stressed" cells are jsut low quality.

There also appears to be a substtial backgroung sigature coming from the
diseased chort.